+++ app/components/comparison/components/comparison.component.ts
@@ -6,6 +6,7 @@ import { ComparisonService } from './comparison.service';
 import { ComparisonCitationService } from './comparison-citation.service';
 import { VersionInformation } from '../../../VersionInformation';
 import { Http } from '@angular/http';
+import { LocalStorageService } from "angular-2-local-storage";
 
 const FileSaver = require('file-saver');
 
@@ -23,7 +24,7 @@ export class ComparisonComponent {
     private ready = false;
     private versionInformation: VersionInformation = new VersionInformation();
     @ViewChild('details') detailsModal: any;
-    private activeRow: Data = new Data(this.http);
+    private activeRow: Data = new Data(this.lss, this.dataServ, this.serv);
     private showTable = false;
     private showTableTooltips = true;
     private tableTooltipsAsFootnotes = false;
@@ -35,7 +36,8 @@ export class ComparisonComponent {
                 public dataServ: ComparisonDataService,
                 public confServ: ComparisonConfigService,
                 public citationServ: ComparisonCitationService,
-                private cd: ChangeDetectorRef) {
+                private cd: ChangeDetectorRef,
+                private lss: LocalStorageService) {
         this.confServ.loadComparison(this.cd);
         this.confServ.loadCriteria(this.cd);
         this.confServ.loadTableData(this.cd);