+++ app/components/output/generic-table/generic-table.component.ts
@@ -1,21 +1,21 @@
 import {
-    Component,
-    Input,
-    Output,
-    EventEmitter,
+    AfterViewChecked,
     ApplicationRef,
     ChangeDetectionStrategy,
-    AfterViewChecked,
     ChangeDetectorRef,
-    OnChanges
-} from "@angular/core";
-import { TableData, Data, CriteriaSelection } from "./../../comparison/shared/index";
-import { ComparisonCitationService } from "./../../comparison/components/comparison-citation.service";
-import { ComparisonConfigService } from "../../comparison/components/comparison-config.service";
-import { DomSanitizer, SafeHtml } from "@angular/platform-browser";
-import { isNullOrUndefined } from "util";
-import { Http } from "@angular/http";
-declare let anchors;
+    Component,
+    EventEmitter,
+    Input,
+    OnChanges,
+    Output
+} from '@angular/core';
+import { CriteriaSelection, Data, TableData } from './../../comparison/shared/index';
+import { ComparisonCitationService } from './../../comparison/components/comparison-citation.service';
+import { ComparisonConfigService } from '../../comparison/components/comparison-config.service';
+import { DomSanitizer, SafeHtml } from '@angular/platform-browser';
+import { Http } from '@angular/http';
+
+declare const anchors;
 
 @Component({
     selector: 'generictable',
@@ -24,30 +24,30 @@ declare let anchors;
     changeDetection: ChangeDetectionStrategy.OnPush
 })
 export class GenericTableComponent implements AfterViewChecked, OnChanges {
-    private counter: number = 0;
+    private counter = 0;
     private table;
 
-    @Input() display: boolean = false;
-    @Input() settings: boolean = false;
-    @Input() columns: Array<TableData> = new Array<TableData>();
+    @Input() display = false;
+    @Input() settings = false;
+    @Input() columns: Array<TableData> = [];
 
-    @Input() data: Array<Data> = new Array<Data>();
-    @Input() query: {[name: string]: CriteriaSelection;} = {};
-    @Input() displayTemplate: boolean = false;
+    @Input() data: Array<Data> = [];
+    @Input() query: { [name: string]: CriteriaSelection; } = {};
+    @Input() displayTemplate = false;
 
     @Input() citationServ: ComparisonCitationService;
 
     @Output() settingsCallback: EventEmitter<any> = new EventEmitter();
     @Output() showDetails: EventEmitter<any> = new EventEmitter();
 
-    @Input() changeNum: number = 0;
+    @Input() changeNum = 0;
 
-    @Input() order: Array<String> = new Array<String>();
+    @Input() order: Array<String> = [];
     @Output() orderChange: EventEmitter<any> = new EventEmitter();
-    @Input() orderOption: Array<number> = new Array<number>();
+    @Input() orderOption: Array<number> = [];
     @Output() orderOptionChange: EventEmitter<any> = new EventEmitter();
 
-    private ctrlCounter: number = 0;
+    private ctrlCounter = 0;
 
     constructor(private ar: ApplicationRef,
                 private confServ: ComparisonConfigService,
@@ -57,21 +57,21 @@ export class GenericTableComponent implements AfterViewChecked, OnChanges {
     }
 
     private orderClick(e: MouseEvent, value: string) {
-        let pos: number = this.order.findIndex(name => name == value);
+        const pos: number = this.order.findIndex(name => name === value);
         if (e.ctrlKey) {
-            this.ctrlCounter = this.order[this.ctrlCounter] == value ? this.ctrlCounter : this.ctrlCounter + 1;
+            this.ctrlCounter = this.order[this.ctrlCounter] === value ? this.ctrlCounter : this.ctrlCounter + 1;
         } else {
             this.ctrlCounter = 0;
         }
-        if (typeof pos != 'undefined' && pos >= 0) {
+        if (typeof pos !== 'undefined' && pos >= 0) {
             this.order[this.ctrlCounter] = value;
-            this.orderOption[this.ctrlCounter] = this.orderOption[pos] == 1 ? -1 : 1;
-            this.orderOption[pos] = pos != this.ctrlCounter ? 0 : this.orderOption[this.ctrlCounter];
+            this.orderOption[this.ctrlCounter] = this.orderOption[pos] === 1 ? -1 : 1;
+            this.orderOption[pos] = pos !== this.ctrlCounter ? 0 : this.orderOption[this.ctrlCounter];
         } else {
             this.order[this.ctrlCounter] = value;
             this.orderOption[this.ctrlCounter] = 1;
         }
-        if (this.ctrlCounter == 0) {
+        if (this.ctrlCounter === 0) {
             for (let i = 1; i < this.orderOption.length; i++) {
                 this.orderOption[i] = 0;
             }
@@ -83,14 +83,14 @@ export class GenericTableComponent implements AfterViewChecked, OnChanges {
 
     private displayOrder(value: string, option: number): boolean {
         if (this.order.length === 0 && this.orderOption.length === 0) {
-            this.order[this.ctrlCounter] = "tag";
+            this.order[this.ctrlCounter] = 'tag';
             this.orderOption[this.ctrlCounter] = 1;
         }
-        return this.order.findIndex(val => val == value) >= 0 && this.orderOption[this.order.findIndex(val => val == value)] == option;
+        return this.order.findIndex(val => val === value) >= 0 && this.orderOption[this.order.findIndex(val => val === value)] === option;
     }
 
     ngAfterViewChecked(): void {
-        this.table = (<any>$("table.table.table-hover"));
+        this.table = (<any>$('table.table.table-hover'));
         this.table.floatThead();
         anchors.options = {
             placement: 'right'
@@ -113,8 +113,8 @@ export class GenericTableComponent implements AfterViewChecked, OnChanges {
             return true;
         }
         let val = true;
-        for (let column of this.confServ.tableDataSet.getTableDataArray()) {
-            if (column.display && data.properties[column.tag] != null && data.properties[column.tag].plain != "") {
+        for (const column of this.confServ.tableDataSet.getTableDataArray()) {
+            if (column.display && data.properties[column.tag] != null && data.properties[column.tag].plain !== '') {
                 return true;
             }
             if (column.display && data.properties[column.tag] != null) {