+++ deploy.sh
@@ -169,7 +169,8 @@ commit+push() {
 
 	disable_expanded_output
 	#--quiet is important here to avoid outputting the repo URL, which may contain a secret token
-	git push --quiet $repo $deploy_branch
+	#--force is important to guarantee a successful push. This overwrites the server history of the branch, thus it should be pushed to a specific branch for deployment only, like gh-pages
+	git push --force --quiet $repo $deploy_branch
 	enable_expanded_output
 }